+++ clox/chunk.h
@@ -21,5 +21,5 @@ typedef struct {
 void initChunk(Chunk *chunk);
 void freeChunk(Chunk *chunk);
 void writeChunk(Chunk *chunk, uint8_t byte, int line);
-int addConstant(Chunk *chunk, Value value);
+size_t addConstant(Chunk *chunk, Value value);
 #endif
